+++ lib/api_callable.js
@@ -329,8 +329,6 @@ exports.NormalApiCaller = NormalApiCaller;
  *   the method call.
  * @return {APICall} func - a bound method on a request stub used
  *   to make an rpc call.
- * @throws - if `settings` has incompatible values, e.g, if bundling
- *   and page_streaming are both configured.
  */
 exports.createApiCall = function createApiCall(
     funcWithAuth, settings, optDescriptor) {
