+++ lib/operations_client.js
@@ -32,7 +32,7 @@ var extend = require('extend');
 var gax = require('./gax');
 extend(gax, require('./api_callable'));
 extend(gax, require('./path_template'));
-extend(gax, require('./page_streaming'));
+extend(gax, require('./paged_iteration'));
 gax.version = require('../package').version;
 
 var SERVICE_ADDRESS = 'longrunning.googleapis.com';
